+++ app/src/main/AndroidManifest.xml
@@ -9,10 +9,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.FindUI">
-        <activity android:name=".ui.ProfileActivity" />
-        <activity android:name=".ui.profile.IntroductionActivity" />
+        <activity android:name=".ui.profile.ProfileActivity" />
+        <activity android:name=".ui.mypage.IntroductionActivity" />
         <activity android:name=".ui.values.ValuesActivity" />
-        <activity android:name=".ui.profile.CreateProfileActivity" />
+        <activity android:name=".ui.mypage.MyPageActivity" />
         <activity android:name=".ui.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
