+++ app/src/main/java/find/ui/ui/mypage/IntroductionActivity.kt
@@ -1,9 +1,10 @@
-package find.ui.ui.profile
+package find.ui.ui.mypage
 
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import find.ui.databinding.ActivityIntroductionBinding
+import find.ui.ui.profile.ProfileViewModel
 
 class IntroductionActivity : AppCompatActivity() {
     private lateinit var binding: ActivityIntroductionBinding
