+++ app/src/main/java/find/ui/ui/profile/ProfileActivity.kt
@@ -1,4 +1,4 @@
-package find.ui.ui
+package find.ui.ui.profile
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
