+++ app/src/main/java/find/ui/ui/profile/ProfileViewModel.kt
@@ -6,13 +6,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class ProfileViewModel : ViewModel() {
-    private val _introLength = MutableLiveData<String>()
+    private val _introLength = MutableLiveData("0/500")
     val introLength: LiveData<String>
         get() = _introLength
 
     val introContent = ObservableField<String>()
-
-    init {
-        _introLength.value = "0/500"
-    }
 }
