+++ app/src/main/res/layout/activity_introduction.xml
@@ -13,7 +13,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.profile.IntroductionActivity">
+        tools:context=".ui.mypage.IntroductionActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/layout_intro_top_menu"
